구장 리뷰 사진 s3 업로드 추가

+++ src/main/java/com/baseballweb/auth/domain/Review.java
@@ -8,7 +8,7 @@ public class Review {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;  // 리뷰 ID
 
-    private int rating;  // 별점
+    private double rating;  // 별점
     private String content;  // 리뷰 내용
     private String imageUrl;  // 사진 URL
 
@@ -24,11 +24,11 @@ public class Review {
         this.id = id;
     }
 
-    public int getRating() {
+    public double getRating() {
         return rating;
     }
 
-    public void setRating(int rating) {
+    public void setRating(double rating) {
         this.rating = rating;
     }
